+++ neoforge/src/main/java/com/lne_archers/neoforge/NeoForgeMod.java
@@ -4,6 +4,7 @@ import com.lne_archers.LNE_ArchersMod;
 import net.minecraft.registry.RegistryKeys;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.neoforge.event.entity.EntityAttributeCreationEvent;
 import net.neoforged.neoforge.registries.RegisterEvent;
 
 @Mod(LNE_ArchersMod.MOD_ID)
@@ -11,6 +12,7 @@ public final class NeoForgeMod {
     public NeoForgeMod(IEventBus modBus) {
         LNE_ArchersMod.init();
         modBus.addListener(RegisterEvent.class, NeoForgeMod::register);
+        modBus.addListener(EntityAttributeCreationEvent.class, NeoForgeMod::registerAttributes);
     }
     public static void register(RegisterEvent event) {
         event.register(RegistryKeys.ITEM, reg -> {
@@ -19,5 +21,8 @@ public final class NeoForgeMod {
         event.register(RegistryKeys.STATUS_EFFECT, reg -> {
             LNE_ArchersMod.registerEffects();
         });
+    }
+    public static void registerAttributes(EntityAttributeCreationEvent event) {
+        LNE_ArchersMod.registerEntityAttributes();
     }
 }